requirements.txt
confluent-kafka[avro]>=2.3.0
jinja2
PyYAML
pytest-cov
coverage
coverage-badge
pre-commit
sphinx>=4.4.0
sphinx-bootstrap-theme>=0.8.1
numpydoc>=1.2
nbsphinx>=0.8.8
nbsphinx-link>=1.3.0
sphinx-copybutton>=0.5.2